[BE] refactor : 전체 폴더 구조 도메인 형으로 변경
