+++ backend/driving-today/src/main/java/com/drivingtoday/domain/instructor/InstructorController.java
@@ -5,7 +5,13 @@ import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import com.drivingtoday.domain.instructor.dto.AvailableInstructorsRequest;
+import com.drivingtoday.domain.instructor.dto.AvailableInstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.RestController;
 
+import java.util.List;
 @RestController
 @RequiredArgsConstructor
 public class InstructorController {
@@ -17,5 +23,17 @@ public class InstructorController {
         InstructorDetailResponse response = instructorFindService.findInstructor(instructorId);
         return ResponseEntity.ok(response);
     }
-}
 
+    @GetMapping("/instructors")
+    @Operation(summary = "예약 가능한 강사 리스트 보기 API")
+    public ResponseEntity<List<AvailableInstructor>> instructorList(@ModelAttribute AvailableInstructorsRequest request) {
+
+        List<Instructor> instructorList = instructorFindService.findAvailableInstructors(request);
+
+        List<AvailableInstructor> availableInstructors = instructorList.stream()
+                .map(AvailableInstructor::from)
+                .toList();
+
+        return ResponseEntity.ok(availableInstructors);
+    }
+}
