+++ backend/driving-today/src/main/java/com/drivingtoday/domain/reservation/ReservationRepository.java
@@ -13,4 +13,3 @@ public interface ReservationRepository extends JpaRepository<Reservation, Long> 
     List<Reservation> findAllByStudentId(Long studentId);
 
 }
-
