+++ backend/driving-today/src/main/java/com/drivingtoday/domain/review/ReviewRepository.java
@@ -1,6 +1,5 @@
 package com.drivingtoday.domain.review;
 
-
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -8,4 +7,3 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface ReviewRepository extends JpaRepository<Review, Long> {
     Page<Review> findAllByInstructorId(Long instructorId, Pageable pageable);
 }
-
