--- backend/driving-today/src/main/java/com/drivingtoday/repository/ReservationRepository.java
@@ -1,6 +1,0 @@
-package com.drivingtoday.repository;
-
-import com.drivingtoday.entity.Reservation;
-import org.springframework.data.jpa.repository.JpaRepository;
-public interface ReservationRepository extends JpaRepository<Reservation, Long> {
-}
