+++ backend/driving-today/src/test/java/com/drivingtoday/domain/instructor/InstructorRepositoryTest.java
@@ -1,8 +1,6 @@
-package com.drivingtoday.repository;
+package com.drivingtoday.domain.instructor;
 
-import com.drivingtoday.entity.Instructor;
 import jakarta.transaction.Transactional;
-import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -10,7 +8,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 @Transactional
